requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
vercel==0.2.1
